Changed support for transformers (now generators)
